change float to double

+++ homework/PC/assignment1/km_openmp.c
@@ -34,8 +34,8 @@ void Find_Distance();
 * @return The number of seconds.
 */
 
-inline float Calculate_Distance(float *x, int i, int j, int Dim, int N_points){
-	float temp=0;
+inline double Calculate_Distance(double *x, int i, int j, int Dim, int N_points){
+	double temp=0;
 	int k;
 	for(k=0; k<Dim; k++){
 		temp+=(x[i+k*N_points]-x[j+k*N_points])*(x[i+k*N_points]-x[j+k*N_points]);
@@ -64,9 +64,9 @@ static inline double monotonic_seconds()
 
 int N_c, N_thread, N_points, Dim, flag;
 int *label, *c_id, *local_id;
-float *x, *sum_dis, *min_c, *local_min;
+double *x, *sum_dis, *min_c, *local_min;
 #if USE_MATRIX
-float **distance_m;
+double **distance_m;
 #endif
 int main(int argc, char** argv){
 	int i,j,num;
@@ -109,8 +109,8 @@ int main(int argc, char** argv){
 //		int N_local=(tid<N_points-num*N_thread)?num+1:num;
 //		printf("tid=%d, N_local=%d\n",tid, N_local);
 
-//		local_min=(float*)malloc(N_c*sizeof(float));
-//		local_id=(int*)malloc(N_c*sizeof(float));
+//		local_min=(double*)malloc(N_c*sizeof(double));
+//		local_id=(int*)malloc(N_c*sizeof(double));
 		Label_Point(tid);
 		while(flag && (i<N_IT)){
 			temp=Find_Medroid(tid);
@@ -191,18 +191,18 @@ char Load_File(char *str){
 }
 
 void Create_Memory(){
-	x=(float*)malloc(N_points*Dim*sizeof(float));
-	sum_dis=(float*)malloc(N_points*sizeof(float));
-	min_c=(float*)malloc(N_c*sizeof(float));
-	local_min=(float*)malloc(N_thread*N_c*sizeof(float));
+	x=(double*)malloc(N_points*Dim*sizeof(double));
+	sum_dis=(double*)malloc(N_points*sizeof(double));
+	min_c=(double*)malloc(N_c*sizeof(double));
+	local_min=(double*)malloc(N_thread*N_c*sizeof(double));
 	label=(int*)malloc(N_points*sizeof(int));
 	c_id=(int*)malloc(N_c*sizeof(int));
 	local_id=(int*)malloc(N_thread*N_c*sizeof(int));
 #if USE_MATRIX
 	int i;
-	distance_m=(float**)malloc(N_points*sizeof(float*));
+	distance_m=(double**)malloc(N_points*sizeof(double*));
 	for(i=0;i<N_points;i++){
-		distance_m[i]=(float*)malloc(N_points*sizeof(float));
+		distance_m[i]=(double*)malloc(N_points*sizeof(double));
 	}
 #endif
 
@@ -225,8 +225,8 @@ void Find_Distance(){
 #endif
 void Label_Point(int tid){
 	int i, j, k;
-	float sum;
-	float temp;
+	double sum;
+	double temp;
 	#pragma omp barrier
 	#pragma omp for // schedule(dynamic)
 	for(i=0; i<N_points; i++){
