+++ homework/PC/assignment1/km_pthread_safe.c
@@ -36,8 +36,8 @@ void Find_Distance(int tid, int num, int offset);
 * @return The number of seconds.
 */
 
-inline float Calculate_Distance(float *x, int i, int j, int Dim, int N_points){
-	float temp=0;
+inline double Calculate_Distance(double *x, int i, int j, int Dim, int N_points){
+	double temp=0;
 	int k;
 	for(k=0; k<Dim; k++){
 		temp+=(x[i+k*N_points]-x[j+k*N_points])*(x[i+k*N_points]-x[j+k*N_points]);
@@ -66,14 +66,14 @@ static inline double monotonic_seconds()
 
 int N_c, N_thread, N_points, Dim, flag;
 int *label, *c_id, *local_id, *info;
-float *x, *sum_dis, *min_c, *local_min;
+double *x, *sum_dis, *min_c, *local_min;
 pthread_t *pthreads;
 pthread_mutex_t mutex1;
 pthread_mutex_t mutex2;
 pthread_barrier_t barrier1;
 pthread_barrier_t barrier2;
 #if USE_MATRIX
-float **distance_m;
+double **distance_m;
 #endif
 
 
@@ -181,10 +181,10 @@ char Load_File(char *str){
 }
 
 void Create_Memory(){
-	x=(float*)malloc(N_points*Dim*sizeof(float));
-	sum_dis=(float*)malloc(N_points*sizeof(float));
-	min_c=(float*)malloc(N_c*sizeof(float));
-	local_min=(float*)malloc(N_thread*N_c*sizeof(float));
+	x=(double*)malloc(N_points*Dim*sizeof(double));
+	sum_dis=(double*)malloc(N_points*sizeof(double));
+	min_c=(double*)malloc(N_c*sizeof(double));
+	local_min=(double*)malloc(N_thread*N_c*sizeof(double));
 	label=(int*)malloc(N_points*sizeof(int));
 	c_id=(int*)malloc(N_c*sizeof(int));
 	local_id=(int*)malloc(N_thread*N_c*sizeof(int));
@@ -193,9 +193,9 @@ void Create_Memory(){
 	
 #if USE_MATRIX
 	int i;
-	distance_m=(float**)malloc(N_points*sizeof(float*));
+	distance_m=(double**)malloc(N_points*sizeof(double*));
 	for(i=0;i<N_points;i++){
-		distance_m[i]=(float*)malloc(N_points*sizeof(float));
+		distance_m[i]=(double*)malloc(N_points*sizeof(double));
 	}
 #endif
 
@@ -220,8 +220,8 @@ void Find_Distance(int tid, int num, int offset){
 
 void Label_Point(int tid, int num, int offset){
 	int i, j, k;
-	float sum;
-	float temp;
+	double sum;
+	double temp;
 //		printf("called label point:\n");
 //	#pragma omp for
 	for(i=offset; i<offset+num; i++){
