+++ homework/PC/assignment3/km_openmp.cpp
@@ -35,8 +35,8 @@ void Find_Distance();
 * @return The number of seconds.
 */
 
-inline float Calculate_Distance(float *x, int i, int j, int Dim, int N_points){
-	float temp=0;
+inline double Calculate_Distance(double *x, int i, int j, int Dim, int N_points){
+	double temp=0;
 	int k;
 	for(k=0; k<Dim; k++){
 		temp+=(x[i+k*N_points]-x[j+k*N_points])*(x[i+k*N_points]-x[j+k*N_points]);
@@ -65,9 +65,9 @@ static inline double monotonic_seconds()
 
 int N_c, N_thread, N_points, Dim, flag, c_id_new;
 int *label, *c_id, *c_old;
-float *x, *sum_dis, *min_c;
+double *x, *sum_dis, *min_c;
 #if USE_MATRIX
-float **distance_m;
+double **distance_m;
 #endif
 int main(int argc, char** argv){
 	int i,j,num;
@@ -267,17 +267,17 @@ char Load_File(char *str){
 }
 
 void Create_Memory(){
-	x=(float*)malloc(N_points*Dim*sizeof(float));
-	sum_dis=(float*)malloc(N_points*sizeof(float));
-	min_c=(float*)malloc(N_c*sizeof(float));
+	x=(double*)malloc(N_points*Dim*sizeof(double));
+	sum_dis=(double*)malloc(N_points*sizeof(double));
+	min_c=(double*)malloc(N_c*sizeof(double));
 	label=(int*)malloc(N_points*sizeof(int));
 	c_id=(int*)malloc(N_c*sizeof(int));
 	c_old=(int*)malloc(N_c*sizeof(int));
 #if USE_MATRIX
 	int i;
-	distance_m=(float**)malloc(N_points*sizeof(float*));
+	distance_m=(double**)malloc(N_points*sizeof(double*));
 	for(i=0;i<N_points;i++){
-		distance_m[i]=(float*)malloc((N_points-i)*sizeof(float));
+		distance_m[i]=(double*)malloc((N_points-i)*sizeof(double));
 	}
 #endif
